# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode:
    enum:
    - AUTH_001
    - AUTH_002
    - AUTH_003
    - AUTH_004
    - AUTH_005
    - AUTH_006
    - AUTH_007
    - USER_001
    - USER_002
    - USER_003
    - USER_004
    - VAL_001
    - VAL_002
    - VAL_003
    - VAL_004
    - DB_001
    - DB_002
    - DB_003
    - DB_004
    - DB_005
    - EXT_001
    - EXT_002
    - EXT_003
    - EXT_004
    - EMAIL_001
    - EMAIL_002
    - MQ_001
    - MQ_002
    - INTERNAL_001
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - CONFLICT
    - BAD_REQUEST
    - UNPROCESSABLE_ENTITY
    type: string
    x-enum-varnames:
    - CodeUnauthorized
    - CodeInvalidCredentials
    - CodeTokenExpired
    - CodeTokenInvalid
    - CodeInsufficientRights
    - CodeAccountLocked
    - CodeAccountNotActivated
    - CodeUserNotFound
    - CodeUserAlreadyExists
    - CodeInvalidUserData
    - CodeEmailNotVerified
    - CodeValidationFailed
    - CodeInvalidInput
    - CodeMissingField
    - CodeInvalidFormat
    - CodeDatabaseError
    - CodeRecordNotFound
    - CodeDuplicateEntry
    - CodeConstraintViolated
    - CodeTransactionFailed
    - CodeServiceUnavailable
    - CodeTimeout
    - CodeRateLimitExceeded
    - CodeCircuitBreakerOpen
    - CodeEmailSendFailed
    - CodeInvalidEmail
    - CodeMessagePublishFailed
    - CodeMessageConsumeFailed
    - CodeInternal
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeConflict
    - CodeBadRequest
    - CodeUnprocessableEntity
  github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail:
    properties:
      code:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode'
      detail:
        type: string
      instance:
        type: string
      meta:
        additionalProperties: true
        type: object
      status:
        type: integer
      time:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
...
  github_com_mr-kaynak_go-core_internal_modules_identity_service.CreateAPIKeyRequest:
    properties:
      expires_at:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      scopes:
        type: string
    required:
    - name
    type: object
...
host: localhost:3000
info:
  contact:
    name: Go-Core Team
    url: https://github.com/mr-kaynak/go-core
  description: Enterprise-grade Go boilerplate with REST API, gRPC services, and notification
    system
  license:
    name: MIT
  title: Go-Core API
  version: 1.0.0
paths:
...
  /api-keys:
    get:
      description: Get all API keys for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: List user's API keys
      tags:
      - API-Keys
    post:
      consumes:
      - application/json
      description: Create a new API key for the authenticated user
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: API key created with raw key
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Create a new API key
      tags:
      - API-Keys
  /api-keys/{id}:
    delete:
      description: Revoke an API key by ID
      parameters:
      - description: API key UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API key revoked
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid API key ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "404":
          description: API key not found
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Revoke an API key
      tags:
      - API-Keys
...
securityDefinitions:
  Bearer:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"