# internal/modules/identity/api/api_key_handler.go
package api

import (
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
	"github.com/mr-kaynak/go-core/internal/core/errors"
	"github.com/mr-kaynak/go-core/internal/core/validation"
	"github.com/mr-kaynak/go-core/internal/modules/identity/service"
)

// APIKeyHandler handles API key related HTTP requests
type APIKeyHandler struct {
	apiKeyService *service.APIKeyService
	auditService  *service.AuditService
}

// NewAPIKeyHandler creates a new API key handler
func NewAPIKeyHandler(apiKeyService *service.APIKeyService) *APIKeyHandler {
	return &APIKeyHandler{
		apiKeyService: apiKeyService,
	}
}

// SetAuditService sets the optional audit service for logging API key events.
func (h *APIKeyHandler) SetAuditService(as *service.AuditService) {
	h.auditService = as
}

// audit is a nil-safe helper that logs an action if audit service is configured.
func (h *APIKeyHandler) audit(c *fiber.Ctx, userID uuid.UUID, action, resourceID string, meta map[string]interface{}) {
	if h.auditService != nil {
		h.auditService.LogAction(&userID, action, "api_key", resourceID, c.IP(), c.Get("User-Agent"), meta)
	}
}

// RegisterRoutes registers API key routes (all require authentication)
func (h *APIKeyHandler) RegisterRoutes(app *fiber.App, authMw fiber.Handler) {
	apiKeys := app.Group("/api/v1/api-keys", authMw)

	apiKeys.Post("/", h.CreateAPIKey)
	apiKeys.Get("/", h.ListAPIKeys)
	apiKeys.Delete("/:id", h.RevokeAPIKey)
}

// CreateAPIKey handles API key creation
// @Summary Create a new API key
// @Description Create a new API key for the authenticated user
// @Tags API-Keys
// @Security Bearer
// @Accept json
// @Produce json
// @Param request body service.CreateAPIKeyRequest true "API key creation request"
// @Success 201 {object} fiber.Map "API key created with raw key"
// @Failure 400 {object} errors.ProblemDetail "Invalid request"
// @Failure 401 {object} errors.ProblemDetail "Not authenticated"
// @Router /api-keys [post]
func (h *APIKeyHandler) CreateAPIKey(c *fiber.Ctx) error {
	userID, ok := c.Locals("userID").(uuid.UUID)
	if !ok {
		return errors.NewUnauthorized("User not authenticated")
	}

	var req service.CreateAPIKeyRequest
	if err := c.BodyParser(&req); err != nil {
		return errors.NewBadRequest("Invalid request body")
	}

	// Validate request
	if err := validation.Struct(req); err != nil {
		return err
	}

	response, err := h.apiKeyService.Create(userID, &req)
	if err != nil {
		return err
	}

	h.audit(c, userID, service.ActionAPIKeyCreated, response.APIKey.ID.String(),
		map[string]interface{}{"key_name": req.Name})
	return c.Status(fiber.StatusCreated).JSON(fiber.Map{
		"message": "API key created successfully. Store the key securely, it will not be shown again.",
		"api_key": response.APIKey,
		"key":     response.RawKey,
	})
}

// ListAPIKeys handles listing a user's API keys
// @Summary List user's API keys
// @Description Get all API keys for the authenticated user
// @Tags API-Keys
// @Security Bearer
// @Produce json
// @Success 200 {object} fiber.Map "List of API keys"
// @Failure 401 {object} errors.ProblemDetail "Not authenticated"
// @Router /api-keys [get]
func (h *APIKeyHandler) ListAPIKeys(c *fiber.Ctx) error {
	userID, ok := c.Locals("userID").(uuid.UUID)
	if !ok {
		return errors.NewUnauthorized("User not authenticated")
	}

	keys, err := h.apiKeyService.List(userID)
	if err != nil {
		return err
	}

	return c.JSON(fiber.Map{
		"api_keys": keys,
	})
}

// RevokeAPIKey handles revoking an API key
// @Summary Revoke an API key
// @Description Revoke an API key by ID
// @Tags API-Keys
// @Security Bearer
// @Produce json
// @Param id path string true "API key UUID"
// @Success 200 {object} fiber.Map "API key revoked"
// @Failure 400 {object} errors.ProblemDetail "Invalid API key ID"
// @Failure 401 {object} errors.ProblemDetail "Not authenticated"
// @Failure 404 {object} errors.ProblemDetail "API key not found"
// @Router /api-keys/{id} [delete]
func (h *APIKeyHandler) RevokeAPIKey(c *fiber.Ctx) error {
	userID, ok := c.Locals("userID").(uuid.UUID)
	if !ok {
		return errors.NewUnauthorized("User not authenticated")
	}

	keyID, err := uuid.Parse(c.Params("id"))
	if err != nil {
		return errors.NewBadRequest("Invalid API key ID format")
	}

	if err := h.apiKeyService.Revoke(keyID, userID); err != nil {
		return err
	}

	h.audit(c, userID, service.ActionAPIKeyRevoked, keyID.String(), nil)
	return c.JSON(fiber.Map{
		"message": "API key revoked successfully",
	})
}